[misc] Added protected attribute _build_info to class Dataset
The Dataset class now has a protected dictionary that keeps track of the
information used to create it, e.g. 'path_to_database', 'queries' ecc.
In this way the variation ChangeDataset can pass only the name of the
new folder and not a whole dataset.
The new dataset will be build out of the info from the current dataset
by simply changing the name and the folder.