[Booking] Add main selection tag in Selection.split() and remove name from __eq__ and __hash__
Importan change: in order to keep track of the main selection after the
split into minimal cuts and weights, we add the tag '___MainSelection'
to the name of the minimal selection generated by Selection.split().

Also the way Selection instantiations are compared is changed: we
require than cuts and weights are exactly the same, but not the name,
since if the same cut comes from a split of different macro selections
it will have for sure a different name.